mirror: initialize a GraphQL database mirror
Summary:
This commit introduces the `Mirror` class that will be the centerpiece
of the persistent-loading API. An instance of this class represents a
mirror of a remote GraphQL database, defined by a particular schema. In
this commit, we add the construction logic, which includes a safety
measure to ensure that the database is used within one version of the
code and schema.

Test Plan:
Unit tests added, with full coverage; run `yarn unit`.

wchargin-branch: mirror-class