STOPSHIP: print Git version on CircleCI
Summary:
We’re getting

        fatal: unknown date format format:%Y%m%d-%H%M

on CircleCI builds. This syntax was introduced in Git 2.6.0, which was
released in August 2015. Perhaps CircleCI has an old version of Git?
(From looking at `v2.5.0:date.c` in Git’s history, this is the error
message that we would expect to see.)

Test Plan:
Push to a branch. Do not merge.

wchargin-branch: stopship-circleci-git-version