travis: cache Yarn (v2)
I have no-op amended the original version of this commit to see how long
it takes to build now that the cache has been set. The original build
was 2m45s of “run” time once the build actually started, which took 30s
or so.

---

This is an experiment. I’m not sure exactly how it works. The docs:

<https://docs.travis-ci.com/user/caching/>

wchargin-branch: travis-cache-yarn